Changed Context.get_value to support integers.
Lua 5.1 does not have integers--all numbers are floats--but as this
would create trouble for us all over the code base, we've decided
to normalize all integer-valued numbers into integers. This means
that the Lua literals "18", "18.", and "18.0" are all returned by
this function as the integer value 18.

Lua 5.2+ has a distinct integer type, meaning that integer values
are returned as integers in any case.